Introduced a scheduler to encapsulate the strategy for distributing work load
